scripts/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.2
pytz>=2023.3
google-cloud-storage>=2.12.0
google-cloud-bigquery>=3.12.0
